Fixed VS build warning.
git-svn-id: https://svnsrv.germanlloyd.org/svn/ExtFEMIO/trunk@170 389f8db8-dd28-11e4-b689-8b400cb78fb4

+++ src/fem/fem_types_int.cpp
@@ -74,7 +74,7 @@ long *entry_type<long>::operator() (const std::string &inp) const {
     std::string msg("boundary condition violated (");
     throw errors::int_error(name, msg + name + ")\n(""" + inp + """)");
   }
-  return new long(value);
+  return new long(static_cast<long>(value));
 }
 
 std::string entry_type<long>::format(const std::unique_ptr<long> &inp) const {
